+++ tests/serde_tests_driver.rs
@@ -1,0 +1,4 @@
+//! Driver for the `serde_tests` subdirectory, which tests support for `serde`.
+
+mod fuzzing;
+mod serde_tests;